% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{biv_quali}
\alias{biv_quali}
\title{Bivariate table for categorical data.}
\usage{
biv_quali(x = NULL, y = NULL, totals = TRUE, useNA = "ifany",
  NA_string = "NA", round_digits = 3, sorting = NULL, latex = FALSE,
  label = NULL, caption = NULL)
}
\arguments{
\item{x}{row variable: a discrete quantitative variable, a
character or a factor}

\item{y}{column variable: a discrete quantitative variable, a
character or a factor}

\item{totals}{print totals?}

\item{useNA}{print NA?}

\item{NA_string}{character used for NA's columns title}

\item{round_digits}{number of rounding digits}

\item{sorting}{sorting can be "\code{asc}" or "\code{desc}", it is
done by row sum}

\item{latex}{output the table using \code{xtable::xtable}}

\item{label}{latex label}

\item{caption}{latex caption}
}
\description{
Bivariate table for categorical data.
}

