% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bivariate_analysis}
\alias{bivariate_analysis}
\title{Dispatcher for (qualitative/quantitative) bivariate tables and
statistics.}
\usage{
bivariate_analysis(x, group, analysis_name, wb)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{group}{a data.frame of 1 column}

\item{analysis_name}{label prefix}

\item{wb}{a WorkBook}
}
\description{
Dispatcher for (qualitative/quantitative) bivariate tables and
statistics.
}

