% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochranq.test.R
\name{cochranq.test}
\alias{cochranq.test}
\title{Cochran q test for dependent proportions}
\source{
\url{https://statisticaconr.blogspot.com/2011/10/cochran-q-test-in-r-nonparametric-test.html} and \url{https://stat.ethz.ch/pipermail/r-help/2006-September/113156.html}
}
\usage{
cochranq.test(mat)
}
\description{
Tipo mc nemar ma per più di due tempi o più di due trattamenti,
nulla che la proporzione sia la stessa.
}
\examples{
mydata <- matrix(c(
                1, 1, 0,
                0, 1, 0,
                1, 1, 1,
                0, 1, 0,
                0, 1, 0,
                0, 1, 1,
                0, 0, 0,
                0, 1, 0,
                1, 1, 0,
                0, 1, 0,
                0, 0, 0,
                0, 0, 1),
             nrow = 12,
             byrow = T,
             dimnames = list(1 : 12,
               c("MarcaA", "MarcaB", "MarcaC")))
cochranq.test(mydata)
}
