% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.backward.R
\name{contr.backward}
\alias{contr.backward}
\title{Backward Difference Coding}
\usage{
contr.backward(k)
}
\arguments{
\item{k}{number of levels of a factor}
}
\description{
mean with confidence interval and some descriptive statistics
}
\examples{
\dontrun{
contr.backward(4)
hsb2 <- read.table('https://stats.idre.ucla.edu/stat/data/hsb2.csv',
                  header = TRUE, sep = ",")

## creating the factor variable race.f
races <- c("Hispanic", "Asian", "African-Am", "Caucasian")
hsb2$race.f <- factor(hsb2$race, labels = races)

## estimates
point_est <- coefficients(summary(lm(write ~ race.f - 1, data = hsb2)))
coef_trt <- coefficients(summary(lm(write ~ race.f, data = hsb2)))
contrasts(hsb2$race.f) <- contr.backward(4)
coef_back <- coefficients(summary(lm(write ~ race.f, data = hsb2)))
diff(point_est[,1])
}

}
\references{
\url{https://stats.oarc.ucla.edu/r/library/r-library-contrast-coding-systems-for-categorical-variables/}
}
