% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.test.data.frame.R
\name{cor.test.data.frame}
\alias{cor.test.data.frame}
\title{cor.test method for data.frame}
\usage{
\method{cor.test}{data.frame}(
  x,
  alternative = c("two.sided", "less", "greater"),
  method = c("pearson", "kendall", "spearman"),
  exact = NULL,
  conf.level = 0.95,
  continuity = FALSE,
  ...
)
}
\arguments{
\item{x}{a data.frame}

\item{alternative}{cor.test alternative parameter}

\item{method}{cor.test method parameter}

\item{exact}{cor.test exact parameter}

\item{conf.level}{cor.test conf.level parameter}

\item{continuity}{cor.test continuity parameter}

\item{...}{further arguments passed to cor.test}
}
\description{
Calculates all coefficients and confidence intervals for the
combinations of variable available in the data.frame
}
\examples{
cor.test(airquality)
}
