% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.test.goodies.R
\name{cor.test_p}
\alias{cor.test_p}
\title{cor.test p-values matrix}
\usage{
cor.test_p(x, ...)
}
\arguments{
\item{x}{matrix/data.frame of variable to be analyzed}

\item{...}{parameters passed to \code{cor.test}}
}
\description{
cor.test p-values matrix
}
