% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{A vector of common descriptive statistics for quantitative data}
\usage{
desc(x, na.rm = TRUE)
}
\arguments{
\item{x}{a quantitative variable}

\item{na.rm}{remove not available values}
}
\description{
A vector of common descriptive statistics for quantitative data
}

