% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat.R
\name{ftest}
\alias{ftest}
\title{F test to compare two linear models (a little utility)}
\usage{
ftest(mod_base = NULL, mod_full = NULL, output = c("str", "df"))
}
\arguments{
\item{mod_base}{if missing, the null model of mod_full is considered}

\item{mod_full}{the model with all the covariates}

\item{output}{str for a string, df for a dataframe}
}
\description{
F test to compare two linear models (a little utility)
}
