% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{gapnc}
\alias{gapnc}
\title{Gap analysis with automatic optimal refitting}
\usage{
gapnc(
  data,
  FUNcluster = kmeans,
  K.max = 10,
  B = 100,
  d.power = 2,
  spaceH0 = "scaledPCA",
  method = "globalSEmax",
  SE.factor = 2,
  ...
)
}
\description{
See exercise2 from hennig
}
