% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci.R
\name{mean_ci}
\alias{mean_ci}
\title{Mean with confidence interval}
\usage{
mean_ci(x, label = "Yes")
}
\arguments{
\item{x}{a vector}
}
\description{
mean with confidence interval and some descriptive statistics
}
