% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{Mode}
\alias{Mode}
\title{mode calculator}
\source{
\url{https://stackoverflow.com/questions/2547402}
}
\usage{
Mode(x)
}
\arguments{
\item{x}{a char or factor variable}
}
\description{
skewness calculator, parametrized as in
\url{http://www.stata.com/manuals13/rsummarize.pdf}
}
