% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mood.median.test.R
\name{mood.median.test}
\alias{mood.median.test}
\title{Mood median test}
\usage{
mood.median.test(x, y)
}
\arguments{
\item{x}{first group of data}

\item{y}{second group of data}
}
\description{
Mood median test
}
