% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_params_ratio.R
\name{n_params_ratio}
\alias{n_params_ratio}
\title{check the 10 n per covariate rules as per harrel}
\usage{
n_params_ratio(x, rule = c("<10", "<15"))
}
\arguments{
\item{x}{a lm, glm(binomial) or coxph object}

\item{rule}{which treshold to check}
}
\description{
check the 10 n per covariate rules as per harrel
}
