% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs2.R
\name{pairs2}
\alias{pairs2}
\title{Battery included scatterplot matrix}
\usage{
pairs2(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{other parameters passed to (but panel, lower.panel,
upper.panel, diag.panel)}
}
\value{

}
\description{
Battery included scatterplot matrix
}
