% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs2.R
\name{pairs_upper}
\alias{pairs_upper}
\title{upper panel for stats::pairs:}
\usage{
pairs_upper(
  x,
  y,
  points_col = lbmisc::col2hex("black", 0.3),
  low_col = "red",
  ...
)
}
\description{
points with lowess estimator
}
