% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_ci.R
\name{prop_ci}
\alias{prop_ci}
\title{Proportion of a label with confidence interval}
\usage{
prop_ci(x, label = "Yes")
}
\arguments{
\item{x}{a vector (factor tipically)}

\item{label}{the label of the desider percentage}
}
\description{
Proportion of a label with confidence interval
}
