% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_ci.R
\name{props_cis}
\alias{props_cis}
\title{Proportions of each label with confidence intervals}
\usage{
props_cis(x)
}
\arguments{
\item{x}{a vector (factor tipically)}
}
\description{
Proportion of a label with confidence interval
}
\examples{
props_cis(gl(2, 10))

}
