% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{quant_density}
\alias{quant_density}
\title{Plot kernel density estimates for quantitative variable of a
list/data.frame}
\usage{
quant_density(x, na.rm = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{x}{a data.frame}

\item{na.rm}{remove NA values from density calculation}

\item{quiet}{if FALSE tells number and names of variables plotted}

\item{...}{further arguments passed to \code{density}}
}
\description{
Plot kernel density estimates for not 0-1 only quantitative
variable of a list/data.frame.
}
\examples{
par(mfrow = c(2, 3))
quant_density(airquality)
}
