% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatter_cor}
\alias{scatter_cor}
\title{Draw a scatterplot with correlation coefficient}
\usage{
scatter_cor(x, y, rho = c("spearman", "pearson"), plot_rho_x = NA,
  plot_rho_y = NA, ...)
}
\arguments{
\item{x}{quantitative variable}

\item{y}{quantitative variable}

\item{rho}{correlation method, either 'spearman' (default) or 'pearson'}

\item{plot_rho_x}{if not NA where to put the correlation estimate string}

\item{plot_rho_y}{if not NA where to put the correlation estimate string}

\item{...}{further parameters passed to plot}
}
\description{
Draw a scatterplot with correlation coefficient
}
