% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroups_forestplotter.R
\name{sg_fplotter_quali}
\alias{sg_fplotter_quali}
\title{Forestplot for subgroup analysis in a dichotomous variable for
survival analysis}
\usage{
sg_fplotter_quali(sg, ...)
}
\description{
Forestplot for subgroup analysis in a dichotomous variable for
survival analysis
}
