% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroups_forestplotter.R
\name{sg_fplotter_quant}
\alias{sg_fplotter_quant}
\title{Forestplot for subgroup analysis in a quantative variable for
survival analysis}
\usage{
sg_fplotter_quant(sg, ...)
}
\description{
Forestplot for subgroup analysis in a quantative variable for
survival analysis
}
