% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{skewness}
\alias{skewness}
\title{skewness calculator}
\usage{
skewness(x, na.rm = FALSE)
}
\arguments{
\item{x}{a quantitative variable}

\item{na.rm}{remove not available values}
}
\description{
skewness calculator, parametrized as in
\url{http://www.stata.com/manuals13/rsummarize.pdf}
}
