% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroups.R
\name{subgroups}
\alias{subgroups}
\title{Function to perform subgroups (effect estimate) analysis}
\usage{
subgroups(
  ep,
  factors_df = NULL,
  subgroups_df = NULL,
  factors_lab = names(factors_df),
  subgroups_lab = names(subgroups_df),
  ...
)
}
\arguments{
\item{ep}{end point (quantitative, factor or Surv).  Estimates with
\code{lm} for quantitative, \code{glm} for binomial factor,
\code{coxph} for \code{Surv}}

\item{factors_df}{a data.frame with main estimate factors (eg in a
trial could be a 1 column with treatment variable)}

\item{subgroups_df}{data.frame of factor for subsetting and
interaction test}

\item{factors_lab}{label for the main factor under study (eg
Treatment in a RCT)}

\item{subgroups_lab}{label for the subgroup (eg could by Histology,
Age class and so on)}
}
\description{
function to perform subgroups analysis given minimal amout of
infos. It should produce a table suitable for forest plotting.
}
