% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{Table}
\alias{Table}
\title{Cross tabulation and table creation}
\usage{
Table(...)
}
\arguments{
\item{...}{Arguments to be passed to table.}
}
\value{
The function return same results of table with NA (if present) and
margins.
}
\description{
This is a wrapper around table (using \code{useNA="ifany"} by
default) and addsmargins.
}

