% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone.R
\name{tableone}
\alias{tableone}
\title{tableone::CreateTableOne wrapper with latex and excel export}
\usage{
tableone(
  vars,
  strata,
  data,
  factorVars,
  includeNA = FALSE,
  test = TRUE,
  testApprox = chisq.test,
  argsApprox = list(correct = TRUE),
  testExact = fisher.test,
  argsExact = list(workspace = 2 * 10^5),
  testNormal = oneway.test,
  argsNormal = list(var.equal = TRUE),
  testNonNormal = kruskal.test,
  argsNonNormal = list(NULL),
  smd = TRUE,
  addOverall = TRUE,
  showAllLevels = TRUE,
  exact = NULL,
  nonnormal = NULL,
  catDigits = 1,
  contDigits = 2,
  pDigits = 3,
  wb = NULL,
  sheet = "tab1",
  print_latex = TRUE,
  label = "table1",
  caption = "Descrittive del campione"
)
}
\arguments{
\item{vars}{same as in tableone::CreateTableOne}

\item{strata}{same as in tableone::CreateTableOne}

\item{data}{same as in tableone::CreateTableOne}

\item{factorVars}{same as in tableone::CreateTableOne}

\item{includeNA}{same as in tableone::CreateTableOne}

\item{test}{same as in tableone::CreateTableOne}

\item{testApprox}{same as in tableone::CreateTableOne}

\item{argsApprox}{same as in tableone::CreateTableOne}

\item{testExact}{same as in tableone::CreateTableOne}

\item{argsExact}{same as in tableone::CreateTableOne}

\item{testNormal}{same as in tableone::CreateTableOne}

\item{argsNormal}{same as in tableone::CreateTableOne}

\item{testNonNormal}{same as in tableone::CreateTableOne}

\item{argsNonNormal}{same as in tableone::CreateTableOne}

\item{smd}{same as in tableone::CreateTableOne}

\item{addOverall}{same as in tableone::CreateTableOne (but TRUE as default)}

\item{showAllLevels}{same as in tableone::print.TableOne (but TRUE as default)}

\item{exact}{same as in tableone::print.TableOne (if NULL
fisher.test is used if lbmisc::fisher_needed)}

\item{nonnormal}{same as in tableone::print.TableOne (if NULL, all
is considered nonnormal)}

\item{catDigits}{same as in tableone::print.TableOne}

\item{contDigits}{same as in tableone::print.TableOne}

\item{pDigits}{same as in tableone::print.TableOne}

\item{wb}{if an openxlsx Workbook is given, Excel exporting to that
one will occurr}

\item{sheet}{sheet name for openxlsx WorkBook}

\item{print_latex}{print to latex}

\item{label}{label for the created latex table}

\item{caption}{caption for the created latex table}
}
\description{
tableone::CreateTableOne wrapper with latex and excel export
}
