% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_univariate.R
\name{univ_mr}
\alias{univ_mr}
\title{A wrapper around univ_perc lightly optimized for multiple response}
\usage{
univ_mr(mr_prefixes = "", ...)
}
\arguments{
\item{mr_prefixes}{the single multiple response}

\item{...}{other parameters passed to univ_perc with exception of
sort_freq (TRUE) and only_non_zero_perc (TRUE)}
}
\description{
A wrapper around univ_perc lightly optimized for multiple response
}
