% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{univ_perc}
\alias{univ_perc}
\title{Percentages table for 0-1 variables}
\usage{
univ_perc(x, sort_freq = TRUE, only_non_zero_perc = TRUE, latex = TRUE,
  latex_placement = "ht", label = NULL, caption = NULL,
  use_comments = TRUE, wb = NULL, sheets = NULL)
}
\arguments{
\item{x}{a (chunk of) data.frame encoding multiple responses (aka
all composed of 0-1 variables)}

\item{sort_freq}{sort by frequencies (descending)?}

\item{only_non_zero_perc}{print only non zero percentages rows}

\item{latex}{output the table using \code{xtable::xtable}}

\item{latex_placement}{table placement for latex printing}

\item{label}{latex label}

\item{caption}{latex caption}

\item{use_comments}{use comments for row (variable) names, if available}

\item{wb}{an openxlsx Workbook; if not NULL the table will be saved
in the workbook too, aside printing}

\item{sheets}{optional sheet names (same length as the number of
tables)}
}
\description{
Previously called \code{univ_mr} (for multiple responses) the function make
descriptive statistics (n, percentages) for 0-1 variables
}
\examples{
set.seed(1)
data <- as.data.frame(matrix(rbinom(100, 1, p = 0.5), ncol = 10))
univ_perc(x = data)
}
