% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{univ_quali}
\alias{univ_quali}
\title{Univariate table for categorical data.}
\usage{
univ_quali(x = NULL, totals = TRUE, useNA = "ifany", NA_string = "NA",
  round_digits = 3, sorting = NULL, latex = FALSE, label = NULL,
  caption = NULL)
}
\arguments{
\item{x}{a discrete quantitative variable, a character or a factor}

\item{totals}{print totals?}

\item{useNA}{print NA?}

\item{NA_string}{character used for NA's columns title}

\item{round_digits}{number of rounding digits}

\item{sorting}{sorting can be "\code{asc}" or "\code{desc}"}

\item{latex}{output the table using xtable::xtable}

\item{label}{latex label}

\item{caption}{latex caption}
}
\description{
Univariate table for categorical data.
}

