% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_table.R
\name{univ_table}
\alias{univ_table}
\title{Univariate table with absolute, relative, cumulative frequencies.}
\usage{
univ_table(var = NULL, totals = TRUE, useNA = "ifany", NA_string = "NA",
  round_digits = 3, sorting = NULL)
}
\arguments{
\item{var}{vector}

\item{totals}{print totals?}

\item{useNA}{print NA?}

\item{NA_string}{character used for NA's columns title}

\item{round_digits}{number of rounding digits}

\item{sorting}{sorting can be "\code{asc} or "\code{desc}"}
}
\value{
A matrix with frequencies.
}
\description{
Univariate table with absolute, relative, cumulative frequencies.
}

