% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_select.R
\name{var_select}
\alias{var_select}
\title{Select predictors of a given}
\usage{
var_select(
  formula,
  data,
  family = c("gaussian", "binomial", "poisson", "multinomial", "cox", "mgaussian"),
  method = c("lasso", "ridge", "step"),
  cv_nfolds = 10,
  lambdas = NULL,
  seed = 1
)
}
\arguments{
\item{formula}{the full model specification}

\item{data}{dataset for search}

\item{family}{type of variable predicted}

\item{method}{which method to use}

\item{cv_nfolds}{crossvalidation subsamples for lasso and ridge}

\item{lambdas}{lambdas for lasso and ridge}

\item{seed}{seed for pseudo-random number generation}
}
\description{
Select best predictors using stepwise AIC, ridge regression or Lasso (with
optimal lambda  determined according to crossvalidation for the latters)
}
