% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_bivariate.R
\name{biv_mr}
\alias{biv_mr}
\title{A wrapper around biv_perc lightly optimized for multiple response}
\usage{
biv_mr(mr_prefixes = "", ...)
}
\arguments{
\item{mr_prefixes}{the single multiple response}

\item{...}{other parameters passed to biv_perc with exception of
sort_freq (TRUE) and only_non_zero_perc (TRUE)}
}
\description{
A wrapper around biv_perc lightly optimized for multiple response
}
