% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{biv_perc}
\alias{biv_perc}
\title{Percentages table for 0-1 variables}
\usage{
biv_perc(x, y, wb = NULL, sheets = NULL)
}
\arguments{
\item{x}{a single qualitative variable or a data.frame
of qualitative variable}

\item{y}{a single qualitative variable or a data.frame
of qualitative variable}

\item{wb}{an openxlsx Workbook; if not NULL the table will be saved
in the workbook too, aside printing}

\item{sheets}{optional sheet names}
}
\description{
Previously called \code{univ_mr} (for multiple responses) the function make
descriptive statistics (n, percentages) for 0-1 variables
}
