% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_bivariate.R
\name{biv_perc}
\alias{biv_perc}
\title{Percentages table for 0-1 variables}
\usage{
biv_perc(
  x = NULL,
  y = NULL,
  xname = NULL,
  yname = NULL,
  collapse_binary_y = FALSE,
  only_non_zero_perc = FALSE,
  style = c("long", "wide"),
  sort_freq = FALSE,
  latex = TRUE,
  latex_placement = "ht",
  label_prefix = "",
  label = "",
  caption_prefix = "",
  caption = "",
  wb = NULL,
  sheets = ""
)
}
\arguments{
\item{x}{a single quantitative variable or a data.frame of
qualitative variable}

\item{y}{a single qualitative variable or a data.frame of
qualitative variable}

\item{collapse_binary_y}{regarding y, keep in row only the 1 in 0-1
variable or the last level for factors}

\item{only_non_zero_perc}{print only non zero percentages rows}

\item{style}{print different groups in different columns}

\item{sort_freq}{decreasing frequencies ordered table}

\item{latex}{output the table using \code{xtable::xtable}}

\item{latex_placement}{table placement for latex printing}

\item{label_prefix}{latex label prefix (useful for a common prefix
in univariate_tables and multivariate_tables). It's ignored if
label is specified (eg the latter is considered
full/complete/correct)}

\item{label}{latex label}

\item{caption_prefix}{a prefix for latex caption (useful for a
common prefix in univariate_tables and multivariate_tables)}

\item{caption}{latex caption}

\item{wb}{an openxlsx Workbook; if not NULL the table will be saved
in the workbook too, aside printing}

\item{sheets}{optional sheet names}
}
\description{
The function make descriptive statistics (n, percentages) for 0-1
(or factor) x variables stratifying for the y variable
}
