% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_bivariate.R
\name{biv_quali}
\alias{biv_quali}
\title{Bivariate table for categorical data.}
\usage{
biv_quali(
  x = NULL,
  y = NULL,
  xname = NULL,
  yname = NULL,
  totals = TRUE,
  tot_row_label = "Tot",
  tot_col_label = "Tot",
  useNA = "ifany",
  perc = c("col", "row", "none"),
  exclude_NA_perc = TRUE,
  NA_string = "NA",
  freq_sorting = c(NA, "desc", "asc"),
  test = c("auto", "none", "fisher", "chisq"),
  test_params = list(),
  latex = TRUE,
  latex_placement = "ht",
  label_prefix = "",
  label = "",
  caption_prefix = "",
  caption = "",
  wb = NULL,
  sheets = ""
)
}
\arguments{
\item{x}{row variable: a discrete quantitative variable, a
character or a factor}

\item{y}{column variable: a discrete quantitative variable, a
character or a factor}

\item{xname}{a string used to identify x variable (used for latex
captions and excel sheets)}

\item{yname}{a string used to identify x variable (used for latex
captions and excel sheets)}

\item{totals}{print totals?}

\item{tot_row_label}{label for rows total}

\item{tot_col_label}{label for columns total}

\item{useNA}{print NA?}

\item{perc}{type of percentages}

\item{exclude_NA_perc}{remove NA from percentages (default = TRUE)}

\item{NA_string}{character used for NA's columns title}

\item{freq_sorting}{freq based sorting: can be \code{NA} (no freq
based sorting) "\code{desc}" (descending) or "\code{asc}"
(ascending). Sorting based on row totals.}

\item{test}{one of 'auto', 'none', 'fisher', 'chi': if \code{auto}
(default) fisher or chi square test will be performed (using
\code{fisher_needed} to decide which one)}

\item{test_params}{a list of parameters to be passed to the test
performing function}

\item{latex}{output the table using \code{xtable::xtable}}

\item{latex_placement}{table placement for latex printing}

\item{label_prefix}{latex label prefix (useful for a common prefix
in univariate_tables and multivariate_tables). It's ignored if
label is specified (eg the latter is considered
full/complete/correct)}

\item{label}{latex label}

\item{caption_prefix}{a prefix for latex caption (useful for a
common prefix in univariate_tables and multivariate_tables)}

\item{caption}{latex caption}

\item{wb}{an openxlsx Workbook; if not NULL the table will be saved
in the workbook too, aside printing}

\item{sheets}{optional sheet names (same length as the number of
tables)}
}
\description{
Bivariate table for categorical data.
}
\examples{
x <- airquality$Ozone < 80
y <- airquality$Month 
biv_quali(x = x, y = y, latex = FALSE, test = 'none')
biv_quali(x = x, y = y, latex = FALSE, test = 'none', freq_sort = 'asc')
biv_quali(x = x, y = y, latex = FALSE, test = 'none', freq_sort = 'desc')
biv_quali(x = x, y = y, latex = FALSE, test = 'none', perc = FALSE )
}
