% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{biv_quant}
\alias{biv_quant}
\title{Bivariate table for a main quantitative vector}
\usage{
biv_quant(x, y, na.rm = FALSE, add_all = TRUE, test = c("none", "anova",
  "kruskal.test"), latex = TRUE, latex_placement = "ht", label = NULL,
  caption = NULL, wb = NULL, sheets = NULL)
}
\arguments{
\item{x}{a quantitative variable}

\item{y}{a discrete quantitative variable, a character or a factor}

\item{na.rm}{exclude missing value group for y}

\item{add_all}{add "All" row}

\item{test}{one of \code{'none'}, \code{'anova'}, or \code{'kruskal.test'}}

\item{latex}{output the table using \code{xtable::xtable}}

\item{latex_placement}{table placement for latex printing}

\item{label}{latex label}

\item{caption}{latex caption}

\item{wb}{an openxlsx Workbook; if not NULL the table will be saved
in the workbook too, aside printing}

\item{sheets}{optional sheet names (same length as the number of
tables)}
}
\description{
Bivariate table for a main quantitative vector
}
