% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{biv_quant}
\alias{biv_quant}
\title{Bivariate table for a main quantitative vector}
\usage{
biv_quant(x, y, na.rm = FALSE, add_all = TRUE, latex = FALSE,
  label = NULL, caption = NULL)
}
\arguments{
\item{x}{a quantitative variable}

\item{y}{a discrete quantitative variable, a character or a factor}

\item{na.rm}{exclude missing value group for y}

\item{add_all}{exclude missing value group for y}

\item{latex}{output the table using \code{xtable::xtable}}

\item{label}{latex label}

\item{caption}{latex caption}
}
\description{
Bivariate table for a main quantitative vector
}

