% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bivariate_tables}
\alias{bivariate_tables}
\title{Dispatcher for (qualitative/quantitative) bivariate tables and
statistics.}
\usage{
bivariate_tables(x, group, analysis_name, wb, quant_test = "none")
}
\arguments{
\item{x}{a \code{data.frame}}

\item{group}{a data.frame of 1 column}

\item{analysis_name}{label prefix}

\item{wb}{a WorkBook}

\item{quant_test}{test for quantitative variables (biv_quant's test param)}
}
\description{
Dispatcher for (qualitative/quantitative) bivariate tables and
statistics.
}

