% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_needed.R
\name{fisher_needed}
\alias{fisher_needed}
\title{Is Fisher's exact test needed}
\usage{
fisher_needed(x = NULL, y = NULL, threshold = 5L, full_report = FALSE)
}
\arguments{
\item{x}{first variable}

\item{y}{second variable}

\item{threshold}{cutoff value under indipendence}

\item{full_report}{return a full report or only the
\code{TRUE}/\code{FALSE} answer}
}
\description{
Check wether Fisher's exact test is needed (or chi-square is ok)
}

