% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_ci.R
\name{median_ci}
\alias{median_ci}
\title{Median bootstrap confidence interval (BCA or parametric for ordered
factors) for numeric and ordered factor}
\usage{
median_ci(x, conf = 0.95, R = 10000, parallel = c("multicore", "no"))
}
\arguments{
\item{x}{data: an integer, numeric or ordered factor}

\item{conf}{confidence.level}

\item{R}{bootstrap repetitions}
}
\description{
Median bootstrap confidence interval (BCA or parametric for ordered
factors) for numeric and ordered factor
}
