% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_ci.R
\name{median_ci}
\alias{median_ci}
\title{Quick median with bootstrap confidence interval for
numeric and ordered factor}
\usage{
median_ci(x, N = 10000)
}
\arguments{
\item{x}{data: an integer, numeric or ordered factor}

\item{N}{bootstrap repetitions}
}
\description{
Quick median with bootstrap confidence interval for
numeric and ordered factor
}
