% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_model.R
\name{pretty_model}
\alias{pretty_model}
\title{pretty print a statistical model}
\usage{
pretty_model(mod, ...)
}
\arguments{
\item{mod}{a lm, glm or coxph object}
}
\description{
export the model as a table with confidence interval and p-values
}
\examples{

## lm
pretty_model(lm(dist ~ speed,  data = cars))

}
