% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{Table}
\alias{Table}
\title{Cross tabulation and table creation}
\usage{
Table(..., useNA = "ifany", f = list(Sum = sum), margin = NULL, quiet = TRUE)
}
\arguments{
\item{...}{Arguments to be passed to table.}

\item{useNA}{display NA counts}

\item{f}{function to be used for summaries}

\item{margin}{addmargins margin parameter}

\item{quiet}{addmargins quiet parameter}
}
\value{
The function return same results of table with NA (if present) and
margins.
}
\description{
This is a wrapper around table (using \code{useNA = "ifany"} by
default) and addsmargins.
}
\examples{
with(airquality, Table(Month, Day))
}
