% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{univ_mr}
\alias{univ_mr}
\title{Percentages table for multiple categorical responses}
\usage{
univ_mr(x, latex = FALSE, label = NULL, caption = NULL)
}
\arguments{
\item{x}{a (chunk of) data.frame encoding multiple responses (aka
all composed of 0-1 variables)}

\item{latex}{output the table using \code{xtable::xtable}}

\item{label}{latex label}

\item{caption}{latex caption}
}
\description{
Percentages table for multiple categorical responses
}

