% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{univ_mr}
\alias{univ_mr}
\title{Percentages table for multiple categorical responses}
\usage{
univ_mr(x, latex = TRUE, label = NULL, caption = NULL, wb = NULL,
  sheets = NULL)
}
\arguments{
\item{x}{a (chunk of) data.frame encoding multiple responses (aka
all composed of 0-1 variables)}

\item{latex}{output the table using \code{xtable::xtable}}

\item{label}{latex label}

\item{caption}{latex caption}

\item{wb}{an openxlsx Workbook; if not NULL the table will be saved
in the workbook too, aside printing}

\item{sheets}{optional sheet names (same length as the number of
tables)}
}
\description{
Percentages table for multiple categorical responses
}

