% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{univ_quali}
\alias{univ_quali}
\title{Univariate table for categorical data.}
\usage{
univ_quali(x = NULL, totals = TRUE, useNA = "ifany", NA_string = "NA",
  freq_sorting = c(NA, "desc", "asc"), latex = TRUE,
  latex_placement = "ht", label = NULL, caption = NULL,
  use_comments = TRUE, wb = NULL, sheets = NULL)
}
\arguments{
\item{x}{a discrete quantitative variable, a character or a factor}

\item{totals}{print totals?}

\item{useNA}{print NA?}

\item{NA_string}{character used for NA's columns title}

\item{freq_sorting}{freq based sorting: can be \code{NA} (no freq
based sorting) "\code{desc}" (descending) or "\code{asc}"
(ascending).}

\item{latex}{output the table using \code{xtable::xtable}}

\item{latex_placement}{table placement for latex printing}

\item{label}{latex label}

\item{caption}{latex caption}

\item{use_comments}{use variable comments for caption if available
(and none specified as comment=)}

\item{wb}{an openxlsx Workbook; if not NULL the table will be saved
in the workbook too, aside printing}

\item{sheets}{optional sheet names (same length as the number of tables)}
}
\description{
Univariate table for categorical data.
}
\examples{

wb = openxlsx::createWorkbook()
univ_quali(x = airquality$Month, wb = wb)
univ_quali(x = airquality[, c('Month')], wb = wb)
univ_quali(x = airquality[, c('Month', 'Day')], wb = wb)
univ_quali(x = airquality[, c('Month', 'Day')], latex = TRUE, wb = wb)
univ_quali(x = airquality[, c('Month', 'Day')], latex = TRUE,
           label = c('tab:airq_month', 'tab:airq_day'),
           caption = c('airquality month', 'airquality day'),
           wb = wb)
univ_quali(list('a' = rep(LETTERS[1:5],2), 'b' = rep(letters[1:5],2)),
           wb = wb)
lbmisc::wb_to_xl(wb = wb, file = '/tmp/univ_quali.xlsx')

}

