% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{univ_quant}
\alias{univ_quant}
\title{Univariate table for quantitative data.}
\usage{
univ_quant(x, latex = FALSE, label = NULL, caption = NULL)
}
\arguments{
\item{x}{a quantitative variable}

\item{latex}{output the table using \code{xtable::xtable}}

\item{label}{latex label}

\item{caption}{latex caption}
}
\description{
Univariate table for quantitative data.
}

