% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{univ_quant}
\alias{univ_quant}
\title{Univariate table for quantitative data.}
\usage{
univ_quant(x, latex = FALSE, label = NULL, caption = NULL, wb = NULL,
  sheets = NULL)
}
\arguments{
\item{x}{a quantitative variable, a data.frame or a list}

\item{latex}{output the table using \code{xtable::xtable}}

\item{label}{latex label}

\item{caption}{latex caption}

\item{wb}{an openxlsx Workbook; if not NULL the table will be saved
in the workbook too, aside printing}

\item{sheets}{optional sheet names (same length as the number of tables)}
}
\description{
Univariate table for quantitative data.
}
\examples{
   univ_quant(x = airquality$Ozone)
   univ_quant(x = airquality[, c('Ozone')])
   univ_quant(x = airquality[, c('Ozone', 'Temp')])
   univ_quant(list('a' = 1:10, 'b' = 2:20))
}

