% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{univ_quant}
\alias{univ_quant}
\title{Univariate table for quantitative data.}
\usage{
univ_quant(x, latex = TRUE, latex_placement = "ht", label = NULL,
  caption = NULL, use_comments = TRUE, wb = NULL, sheets = NULL)
}
\arguments{
\item{x}{a quantitative variable, a data.frame or a list}

\item{latex}{output the table using \code{xtable::xtable}}

\item{latex_placement}{table placement for latex printing}

\item{label}{latex label}

\item{caption}{latex caption}

\item{use_comments}{use comments for row (variable) names, if available}

\item{wb}{an openxlsx Workbook; if not NULL the table will be saved
in the workbook too, aside printing}

\item{sheets}{optional sheet names (same length as the number of tables)}
}
\description{
Univariate table for quantitative data.
}
\examples{

wb = openxlsx::createWorkbook()
univ_quant(x = airquality$Ozone, sheet = 'ozone', wb = wb)
univ_quant(x = airquality[, c('Ozone')], wb = wb)
univ_quant(x = airquality[, c('Ozone', 'Temp')], wb = wb)
univ_quant(list('a' = 1:10, 'b' = 2:20), wb = wb)
lbmisc::wb_to_xlsx(wb = wb, file = '/tmp/univ_quant.xlsx')

}

