% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{univariate_analysis}
\alias{univariate_analysis}
\title{Dispatcher for (qualitative/quantitative) univariate tables and
statistics.}
\usage{
univariate_analysis(x, wb = NULL)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{wb}{a WorkBook}
}
\description{
Dispatcher for (qualitative/quantitative) univariate tables and
statistics.
}

