% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{univariate_tables}
\alias{univariate_tables}
\title{Dispatcher for (qualitative/quantitative) univariate tables and
statistics.}
\usage{
univariate_tables(x, wb = NULL, latex = TRUE, exclude = NULL,
  style = c("raw", "type"), date_preproc = function(d) factor(format(d,
  "\%Y-\%m")), mr_prefixes = NULL, univ_perc_params = list(),
  univ_quant_params = list(), univ_quali_params = list())
}
\arguments{
\item{x}{a \code{data.frame}}

\item{wb}{a WorkBook}

\item{latex}{use latex for printing (default = TRUE)}

\item{exclude}{character vector of data.frame name to be ignored}

\item{style}{if 'raw' display variables in the order given, if
'type' display variable by type}

\item{date_preproc}{function to preprocess Date columns. By default
make them Year-month factors; currently only factor, numerics
and 0-1 variable are not ignored, therefore date_preproc should
be a coercer to those types}

\item{mr_prefixes}{variable name prefixes (character vector); each
of these identify a set of variable who belong to the same
multiple response question}

\item{univ_perc_params}{other options (named list) for univ_perc}

\item{univ_quant_params}{other options (named list) for univ_quant}

\item{univ_quali_params}{other options (named list) for univ_quali}
}
\description{
Dispatcher for (qualitative/quantitative) univariate tables and
statistics.
}
\examples{

wb = openxlsx::createWorkbook()
univariate_tables(airquality, wb = wb)
lbmisc::wb_to_xl(wb = wb, file = '/tmp/univariate_tables.xlsx')

}
